+++ provinder_8081/src/main/java/com/grace/it/service/DeptManagerService.java
@@ -20,7 +20,7 @@ public interface DeptManagerService {
 
     public Integer updateByCondition(DeptManager deptManager);
 
-    public Long insert(DeptManager deptManager);
+    public Integer insert(DeptManager deptManager);
 
     public Integer deleteByEmpNo(Long empNo);
 }
