+++ provinder_8081/src/main/java/com/grace/it/service/EmployeeService.java
@@ -21,7 +21,7 @@ public interface EmployeeService {
 
     public Integer updateByCondition(Employee employee);
 
-    public Long insert(Employee employee);
+    public Integer insert(Employee employee);
 
     public Integer deleteByEmpNo(Long empNo);
 }
